requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
requests>=2.31.0
chromadb>=0.4.0
cohere>=5.0.0
google-genai>=1.0.0
google-cloud-speech>=2.20.0
